Made shots

+++ model/State.go
@@ -1,8 +1,7 @@
 package model
 
-
 type State struct {
-	Timestamp int
+	Timestamp int64
 	Players   []Player
 	Zombies   []Zombie
 	Shoots    []Shoot
